+++ CHANGELOG.markdown
@@ -1,3 +1,8 @@
+4.5
+---
+* Provide access to the typename in lensRules naming function.
+* `makeFields` camelcasing rules now properly support types with camelcasing. `MyType` with field `myTypeFieldA` generates `fieldA` now. Previously the prefix ignore capitalization and the field would need to be named `mytypeFieldA`.
+
 4.4.0.2
 ---
 * `text` 1.2.0.0 support